Common: Rename register{,_agent_event_serializers}.py
