+++ monkey/common/agent_event_serializers/__init__.py
@@ -1,2 +1,2 @@
 from .agent_event_serializer_registry import AgentEventSerializerRegistry
-from .register import register_common_agent_event_serializers
+from .register_agent_event_serializers import register_common_agent_event_serializers
